soc: qcom: watchdog_v2: use kmemleak_not_leak for cpu context dump data
cpu_data & cpu_buf are virtual addresses,which are
converted to their physical addresses by using
virt_to_phys() API. Then this is added to the mem dump
apps data table by calling msm_dump_table_register().
This memory is not intended to be freed. When kmemleak
scans, it can't deduce that the reference we are
holding is via physical address and not virtual address.
Hence, kmemleak will treat this as a leak as it can't
find a reference to this variable. Supress this by
using kmemleak_not_leak.

Change-Id: I4ee30c87a29d66f34a5c8a6c18f2ebc59760e981
Signed-off-by: Kishan Kumar <kishank@codeaurora.org>
Signed-off-by: W4TCH0UT <ateekujjawal@gmail.com>