msm: qpnp-power-on: add regulator support for PON device
There is a requirement for configuring certain spare register
bits of PON peripheral. Since there is a chance for multiple
clients to configure these bits, a regulator device can handle
such configuration easily. Add support for it.

Regulator configuration requires an address inside PON peripheral
and a bit position which needs to be controlled to be specified
in the device tree subnode.

CRs-Fixed: 840571
Change-Id: I62d0c3f49e777103e102229353faaf2e8ee1fad6
Signed-off-by: Subbaraman Narayanamurthy <subbaram@codeaurora.org>
Signed-off-by: W4TCH0UT <ateekujjawal@gmail.com>