msm: vidc: Generalize decoder DCVS threshold for all usecases
Previously the lower DCVS threshold assumed that the client (typically)
display would hold around 5 buffers.  However, this really isn't the
case for certain clients which might hold additional buffers for
arbitrary reasons.  This breaks the assumptions and ultimately ruins the
DCVS algorithm.

To avoid this, set the lower threshold to mean all extra buffers that
are allocated outside the minimum count that the firmware requires (plus
one to account for buffers in transit across various components).

This commit also has parts of "5dea60 msm: vidc: Update loads for DCVS"
squashed in.

Change-Id: Id85305243bd36ea4049bd9457d07b3d00729196b
Signed-off-by: Deva Ramasubramanian <dramasub@codeaurora.org>
Signed-off-by: W4TCH0UT <ateekujjawal@gmail.com>