ASoC: msm: qdsp6v2: Track no wait commands
Add a command queue for no wait commands. Use queue
to track all active no wait commands and remove
them from the queue when they have correctly
returned. If a queue is not used, a wait command
could be improperly recognized as a no wait
command and cause a timeout.

Change-Id: I0a5033b6942418dc746d64e34b3da0df50c5fc7c
Signed-off-by: Ben Romberger <bromberg@codeaurora.org>
Signed-off-by: W4TCH0UT <ateekujjawal@gmail.com>