cnss: Fix a boundary check bug in cnss prealloc driver
Currently the driver gives a memory slot only if its size is greater
than the memory requested by the caller. This will waste a bigger slot
if the memory requested is equal to the slot size. Hence fix it by
adding the equal check when a memory slot is requested.

Change-Id: Ib48477607e9332c8942894301ead606a31fa3284
CRs-Fixed: 978073
Signed-off-by: Yue Ma <yuem@codeaurora.org>
Signed-off-by: Komal Seelam <kseelam@codeaurora.org>
Signed-off-by: Sarada Prasanna Garnayak <sgarna@codeaurora.org>
Signed-off-by: W4TCH0UT <ateekujjawal@gmail.com>